% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_plot_umap.R
\name{plot_and_save_umap}
\alias{plot_and_save_umap}
\title{Plot and Save UMAP Clusters}
\usage{
plot_and_save_umap(cotan_obj, out_dir)
}
\arguments{
\item{cotan_obj}{A \code{COTAN} object with clustering complete.}

\item{out_dir}{Directory to save the output PDF.}
}
\description{
Plot and Save UMAP Clusters
}
