% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_plot_clean.R
\name{plot_cleaned_cotan}
\alias{plot_cleaned_cotan}
\title{Plot and Save Cleaned COTAN Plots}
\usage{
plot_cleaned_cotan(
  cotan_obj,
  output_dir = NULL,
  prefix = "cotan_clean",
  include_pca = TRUE
)
}
\arguments{
\item{cotan_obj}{A cleaned \code{COTAN} object.}

\item{output_dir}{Character. Optional directory to save the plots as PDF files. Default is \code{NULL}.}

\item{prefix}{Character. Prefix for the saved plot filenames. Default is \code{"cotan_clean"}.}

\item{include_pca}{Boolean. Whether to calculate the PCA associated with the normalized matrix. Default is \code{TRUE}.}
}
\value{
A list of plots/data returned by \code{cleanPlots}.
}
\description{
Generates diagnostic plots for a cleaned COTAN object using the \code{cleanPlots} function.
These plots include PCA of cells, cells' UDE against PCA, and nu plots.
It can optionally save these plots to a specified directory.
}
