% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_plot_coex.R
\name{plot_coex_heatmap}
\alias{plot_coex_heatmap}
\title{Plot and Save COEX Heatmap}
\usage{
plot_coex_heatmap(
  cotan_obj,
  genes_list = NULL,
  sample_size = 10L,
  output_dir = NULL,
  prefix = "cotan_coex",
  p_value_threshold = 0.01,
  cores = 1L
)
}
\arguments{
\item{cotan_obj}{A \code{COTAN} object with a calculated COEX matrix.}

\item{genes_list}{A named list of genes (transcripts) to plot. Default is \code{NULL}.}

\item{sample_size}{Integer. Number of features to sample if \code{genes_list} is \code{NULL}. Default is \code{10L}.}

\item{output_dir}{Character. Optional directory to save the heatmap as a PDF file. Default is \code{NULL}.}

\item{prefix}{Character. Prefix for the saved plot filename. Default is \code{"cotan_coex"}.}

\item{p_value_threshold}{Numeric. The p-value threshold. Default is 0.01.}

\item{cores}{Integer. Number of cores to use. Default is 1L.}
}
\value{
A \code{ggplot2} object representing the heatmap.
}
\description{
Generates a COEX heatmap for selected genes list using the \code{heatmapPlot} function.
It can optionally save the heatmap to a specified directory.
}
