% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot_gdi.R
\name{plot_gdi}
\alias{plot_gdi}
\title{Plot and Save GDI Distribution}
\usage{
plot_gdi(
  cotan_obj,
  out_dir,
  genes = NULL,
  sample_size = 10L,
  condition = "",
  stat_type = "S",
  gdi_threshold = 1.43,
  gdi_in = NULL
)
}
\arguments{
\item{cotan_obj}{A \code{COTAN} object.}

\item{out_dir}{Character. Directory to save the output PDF.}

\item{genes}{Named list. A list of genes to label, where each array will have a different color.
Default is \code{NULL}.}

\item{sample_size}{Integer. Number of features to sample if \code{genes} is \code{NULL}. Default is \code{100L}.}

\item{condition}{Character. A string corresponding to the condition/sample (used only for the title).
Default is \code{""}.}

\item{stat_type}{Character. Type of statistic to be used ("S" for Pearson's chi-squared, "G" for G-test).
Default is \code{"S"}.}

\item{gdi_threshold}{Numeric. The threshold level used in a SimpleGDIUniformityCheck. Default is \code{1.43}.}

\item{gdi_in}{Data.frame. A pre-calculated GDI data frame to speed up the process. Default is \code{NULL}.}
}
\value{
Invisible character string containing the path to the saved PDF.
}
\description{
Generates and saves the Global Differentiation Index (GDI) plot.
Automatically handles the mandatory 'genes' argument by sampling a specified
number of features if none are provided.
}
